+++ test/__init__.py
@@ -3,6 +3,7 @@
 """Utility functions used for testing. """
 
 
+import json
 import os
 import os.path as op
 import contextlib
@@ -10,18 +11,20 @@ import threading
 import multiprocessing as mp
 import functools as ft
 import sys
+import time
 import re
-
 
 
 # py3
 try:
+    import queue
     import http.server as http
     from io import StringIO
     from unittest import mock
 
 # py2
 except ImportError:
+    import Queue as queue
     from StringIO import StringIO
     import SimpleHTTPServer as http
     http.HTTPServer = http.BaseHTTPServer.HTTPServer
@@ -52,20 +55,55 @@ def indir(dir):
 
 
 class HTTPServer(mp.Process):
-    """Simple HTTP server which serves files from a specified directory.
+    """Simple HTTP server which serves files from a specified directory, and
+    can accept JSON data via POST requests.
 
     Intended to be used via the :func:`server` context manager function.
     """
+
+
+    class Handler(http.SimpleHTTPRequestHandler):
+        def __init__(self, *args, **kwargs):
+            self.posts = kwargs.pop('posts', queue.Queue())
+            http.SimpleHTTPRequestHandler.__init__(self, *args, **kwargs)
+
+        @classmethod
+        def ctr(cls, posts):
+            return ft.partial(cls, posts=posts)
+
+        def do_POST(self):
+
+            nbytes = int(self.headers['Content-Length'])
+            data   = json.loads(self.rfile.read(nbytes).decode())
+
+            self.posts.put(data)
+            self.send_response(200)
+            self.send_header('Content-type', 'text/html')
+            self.end_headers()
+            self.wfile.write(b'OK')
+
     def __init__(self, rootdir):
+
         mp.Process.__init__(self)
         self.daemon = True
         self.rootdir = rootdir
-        handler = http.SimpleHTTPRequestHandler
-        self.server = http.HTTPServer(('', 0), handler)
+        self.__postq = mp.Queue()
+        self.__posts = []
+        handler = HTTPServer.Handler.ctr(self.__postq)
+        self.server = http.HTTPServer(('0.0.0.0', 0), handler)
         self.shutdown = mp.Event()
 
     def stop(self):
         self.shutdown.set()
+
+    @property
+    def posts(self):
+        while True:
+            try:
+                self.__posts.append(self.__postq.get_nowait())
+            except queue.Empty:
+                break
+        return list(self.__posts)
 
     @property
     def port(self):
@@ -84,6 +122,12 @@ def server(rootdir=None):
     ``rootdir`` (defaults to the current working directory), then shut it down
     afterwards.
     """
+    # pause for a bit to allow OS to free
+    # resources (in case we are calling
+    # server() multiple times in quick
+    # succession)
+    time.sleep(3)
+
     if rootdir is None:
         rootdir = os.getcwd()
     srv = HTTPServer(rootdir)
@@ -141,11 +185,15 @@ def mock_input(*responses):
     is a callable, in which case it is called, and then the next non-callable
     response returned. This gives us a hacky way to manipulate things while
     stuck in an input REPL loop.
+
+    An error is raised if input is not called the expected number of times
     """
 
     resp = iter(responses)
+    count = [0]
 
     def _input(*a, **kwa):
+        count[0] += 1
         n = next(resp)
         while callable(n):
             n()
@@ -157,6 +205,9 @@ def mock_input(*responses):
 
     with mock.patch(target, _input):
         yield
+
+    if count[0] != len(responses):
+        raise AssertionError('Expected number of inputs not provided')
 
 
 def strip_ansi_escape_sequences(text):