+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # FSL installer script release history
 
 
+# 3.7.0 (Monday 29th January 2024)
+
+ - FSL installations are now registered with a remote server. Basic
+   installation and system information is sent as part of the registration
+   process. This can be skipped by passing the `--skip_registration` / `-r`
+   command-line option.
+ - Add another progress reporting mechanism.
+
+
 # 3.6.0 (Thursday 18th January 2024)
 
  - Add a new progress reporting mechanism.